+++ datanode/gateway/graphql/resolvers.go
@@ -271,10 +271,6 @@ func (r *VegaResolverRoot) RankingScore() RankingScoreResolver {
 	return (*rankingScoreResolver)(r)
 }
 
-func (r *VegaResolverRoot) RewardScore() RewardScoreResolver {
-	return (*rewardScoreResolver)(r)
-}
-
 func (r *VegaResolverRoot) KeyRotation() KeyRotationResolver {
 	return (*keyRotationResolver)(r)
 }
@@ -1060,7 +1056,7 @@ func (r *myQueryResolver) HistoricBalances(ctx context.Context, filter *v2.Accou
 	gb := make([]v2.AccountField, len(groupBy))
 	for i, g := range groupBy {
 		if g == nil {
-			return nil, fmt.Errorf("Nil group by")
+			return nil, errors.New("nil group by")
 		}
 		gb[i] = *g
 	}